+++ test/borrower/shelf.t.sol
@@ -13,7 +13,6 @@ import "./mock/subscriber.sol";
 import "./mock/assessor.sol";
 import "../lender/mock/reserve.sol";
 
-
 contract ShelfTest is Test {
     Shelf shelf;
     NFTMock nft;
@@ -35,29 +34,27 @@ contract ShelfTest is Test {
         shelf = new Shelf(address(currency), address(title), address(pile), address(ceiling));
         shelf.depend("reserve", address(reserve));
         shelf.depend("assessor", address(assessor));
-
     }
 
-    function _issue(uint256 tokenId_, uint loan_) internal {
+    function _issue(uint256 tokenId_, uint256 loan_) internal {
         title.setReturn("issue", loan_);
         title.setReturn("ownerOf", address(this));
 
-        uint loanId = shelf.issue(address(nft), tokenId_);
+        uint256 loanId = shelf.issue(address(nft), tokenId_);
         assertEq(loanId, loan_);
         assertEq(shelf.nftlookup(keccak256(abi.encodePacked(address(nft), tokenId_))), loan_);
     }
 
-    function _lock(uint256 tokenId_, uint loan_) internal {
+    function _lock(uint256 tokenId_, uint256 loan_) internal {
         shelf.lock(loan_);
 
         assertEq(nft.calls("transferFrom"), 1);
         assertEq(nft.values_address("transferFrom_to"), address(shelf));
         assertEq(nft.values_address("transferFrom_from"), address(this));
         assertEq(nft.values_uint("transferFrom_tokenId"), tokenId_);
-
     }
 
-    function _borrow(uint loan_, uint currencyAmount_) internal {
+    function _borrow(uint256 loan_, uint256 currencyAmount_) internal {
         shelf.borrow(loan_, currencyAmount_);
         assertEq(currency.calls("transferFrom"), 1);
 
@@ -65,27 +62,27 @@ contract ShelfTest is Test {
         assertEq(pile.calls("accrue"), 1);
         assertEq(pile.calls("incDebt"), 1);
         assertEq(shelf.balance(), currencyAmount_);
-        uint loanBalance = shelf.balances(loan_);
+        uint256 loanBalance = shelf.balances(loan_);
         assertEq(loanBalance, currencyAmount_);
     }
 
-    function _withdraw(uint loan_, uint currencyAmount_) internal {
-        uint totalBalance = shelf.balance();
-        uint loanBalance = shelf.balances(loan_);
+    function _withdraw(uint256 loan_, uint256 currencyAmount_) internal {
+        uint256 totalBalance = shelf.balance();
+        uint256 loanBalance = shelf.balances(loan_);
         assertEq(totalBalance, currencyAmount_);
         assertEq(loanBalance, currencyAmount_);
         assertEq(pile.values_uint("incDebt_currencyAmount"), currencyAmount_);
 
         shelf.withdraw(loan_, currencyAmount_, address(this));
-        assertEq(totalBalance-currencyAmount_, shelf.balance());
-        assertEq(loanBalance-currencyAmount_, shelf.balances(loan_));
+        assertEq(totalBalance - currencyAmount_, shelf.balance());
+        assertEq(loanBalance - currencyAmount_, shelf.balances(loan_));
         assertEq(currency.calls("transferFrom"), 2);
         assertEq(currency.values_address("transferFrom_from"), address(shelf));
         assertEq(currency.values_address("transferFrom_to"), address(this));
         assertEq(currency.values_uint("transferFrom_amount"), currencyAmount_);
     }
 
-    function _repay(uint loan_, uint currencyAmount_) internal {
+    function _repay(uint256 loan_, uint256 currencyAmount_) internal {
         pile.setReturn("debt_loan", currencyAmount_);
         shelf.repay(loan_, currencyAmount_);
 
@@ -101,7 +98,7 @@ contract ShelfTest is Test {
         assertEq(currency.values_uint("transferFrom_amount"), currencyAmount_);
     }
 
-    function _recover(uint loan_, address usr_, uint currencyAmount_, uint debt_) internal {
+    function _recover(uint256 loan_, address usr_, uint256 currencyAmount_, uint256 debt_) internal {
         pile.setReturn("debt_loan", debt_);
         shelf.recover(loan_, usr_, currencyAmount_);
         assertEq(pile.calls("accrue"), 2);
@@ -113,8 +110,8 @@ contract ShelfTest is Test {
         assertEq(currency.values_uint("transferFrom_amount"), currencyAmount_);
     }
 
-    uint loan  = 1;
-    uint currencyAmount = 100;
+    uint256 loan = 1;
+    uint256 currencyAmount = 100;
     uint256 tokenId = 55;
 
     function testBorrow() public {
@@ -130,7 +127,7 @@ contract ShelfTest is Test {
 
     function testFailBorrowNFTNotLocked() public {
         nft.setReturn("ownerOf", address(this));
-        _issue(tokenId,loan);
+        _issue(tokenId, loan);
         _borrow(loan, currencyAmount);
     }
 
@@ -142,7 +139,7 @@ contract ShelfTest is Test {
 
     function testFailWithdrawNFTNotLocked() public {
         nft.setReturn("ownerOf", address(this));
-        _issue(tokenId,loan);
+        _issue(tokenId, loan);
         _borrow(loan, currencyAmount);
         shelf.claim(loan, address(1));
         _withdraw(loan, currencyAmount);
@@ -164,12 +161,12 @@ contract ShelfTest is Test {
         testLock();
         _borrow(loan, currencyAmount);
         _withdraw(loan, currencyAmount);
-        _recover(loan, address(1), currencyAmount-10, currencyAmount);
+        _recover(loan, address(1), currencyAmount - 10, currencyAmount);
     }
 
     function testFailRepayNFTNotLocked() public {
         nft.setReturn("ownerOf", address(this));
-        _issue(tokenId,loan);
+        _issue(tokenId, loan);
         _borrow(loan, currencyAmount);
         _withdraw(loan, currencyAmount);
         shelf.claim(loan, address(1));
@@ -188,11 +185,11 @@ contract ShelfTest is Test {
 
     function testIssue() public {
         nft.setReturn("ownerOf", address(this));
-       _issue(tokenId, loan);
+        _issue(tokenId, loan);
     }
 
     function testMultiple_Issue() public {
-        uint secondLoan = 2;
+        uint256 secondLoan = 2;
         nft.setReturn("ownerOf", address(this));
 
         _issue(tokenId, loan);
@@ -206,14 +203,14 @@ contract ShelfTest is Test {
         title.setReturn("issue", secondLoan);
         title.setReturn("ownerOf", address(this));
 
-        uint loanId = shelf.issue(address(nft), tokenId);
+        uint256 loanId = shelf.issue(address(nft), tokenId);
         assertEq(loanId, secondLoan);
         nft.setReturn("ownerOf", secondLoan);
         assertEq(shelf.nftlookup(keccak256(abi.encodePacked(address(nft), tokenId))), secondLoan);
     }
 
     function testFailMultiple_Issue() public {
-        uint secondLoan = 2;
+        uint256 secondLoan = 2;
         nft.setReturn("ownerOf", address(this));
 
         _issue(tokenId, loan);
