+++ test/mock/mock.sol
@@ -1,44 +1,45 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity >=0.7.6;
+
 import "forge-std/Test.sol";
 import "tinlake-math/math.sol";
 
 interface SimpleTokenLike {
-    function balanceOf(address) external view returns (uint);
-    function transferFrom(address, address, uint) external returns (bool);
-    function mint(address, uint) external;
-    function burn(address, uint) external;
-    function totalSupply() external view returns (uint);
-    function approve(address usr, uint amount) external;
+    function balanceOf(address) external view returns (uint256);
+    function transferFrom(address, address, uint256) external returns (bool);
+    function mint(address, uint256) external;
+    function burn(address, uint256) external;
+    function totalSupply() external view returns (uint256);
+    function approve(address usr, uint256 amount) external;
 }
 
 contract Mock is Test, Math {
     // counting calls
-    mapping (bytes32 => uint) public calls;
+    mapping(bytes32 => uint256) public calls;
 
     // returns
-    mapping (bytes32 => uint) public values_return;
-    mapping (bytes32 => address) public values_address_return;
-    mapping (bytes32 => bool) public values_bool_return;
-    mapping (bytes32 => int) public values_int_return;
-    mapping (bytes32 => bytes32) public values_bytes32_return;
+    mapping(bytes32 => uint256) public values_return;
+    mapping(bytes32 => address) public values_address_return;
+    mapping(bytes32 => bool) public values_bool_return;
+    mapping(bytes32 => int256) public values_int_return;
+    mapping(bytes32 => bytes32) public values_bytes32_return;
 
     // passed parameter
-    mapping (bytes32 => uint) public values_uint;
-    mapping (bytes32 => address) public values_address;
-    mapping (bytes32 => bytes32) public values_bytes32;
+    mapping(bytes32 => uint256) public values_uint;
+    mapping(bytes32 => address) public values_address;
+    mapping(bytes32 => bytes32) public values_bytes32;
 
-    mapping (bytes32 =>bool) method_fail;
+    mapping(bytes32 => bool) method_fail;
 
     // function values(bytes32 name) public returns (uint) {return values_uint[name];}
     // function values(bytes32 name) public returns (address) {return values_address[name];}
 
-    function call(bytes32 name) internal returns (uint) {
+    function call(bytes32 name) internal returns (uint256) {
         calls[name]++;
         return values_return[name];
     }
 
-    function setReturn(bytes32 name, uint returnValue) public {
+    function setReturn(bytes32 name, uint256 returnValue) public {
         values_return[name] = returnValue;
     }
 
@@ -50,16 +51,16 @@ contract Mock is Test, Math {
         values_bytes32_return[name] = returnValue;
     }
 
-    function setIntReturn(bytes32 name, int returnValue) public {
+    function setIntReturn(bytes32 name, int256 returnValue) public {
         values_int_return[name] = returnValue;
     }
 
-    function setReturn(bytes32 name, bool flag, uint value) public {
+    function setReturn(bytes32 name, bool flag, uint256 value) public {
         setReturn(name, flag);
         setReturn(name, value);
     }
 
-    function setReturn(bytes32 name, address addr, uint value) public {
+    function setReturn(bytes32 name, address addr, uint256 value) public {
         setReturn(name, addr);
         setReturn(name, value);
     }
